# Improve semantic syntax highlighting