- Correctly use `save-excursion` before `save-restriction` macro
 - Remove unnecessary `buffer` param from `idris-highlight-input-region`
 - Use `pcase-dolist` instead of `cl-loop & pcase`
 - Per buffer size controlled semantic source highlighting 
 - Decoupling debug mode for semantic highlighting from the default happy path

**Benchmark:**
Run twice with 5 and 10 repetitions
> The (x y z) represent list of the total elapsed time for execution, the number of garbage collections that ran, and the time taken by garbage collection.

**main** (before change)
;; 5 (4.345052 10 0.3808530000000001)
;; 10 (9.112914 21 0.8207839999999997)

**impro-sem-high** (after change)
;; idris-semantic-source-highlighting-max-buffer-size 32768
**;; 5 (4.123512 8 0.37952299999999894)
;; 10 (8.677349000000001 17 0.8196910000000006)**

;; idris-semantic-source-highlighting-max-buffer-size 9000 (snake/Main.idr loaded without highlighting)
;; 5 (3.922993 4 0.20145000000000124)
;; 10 (7.300091 9 0.45280699999999996)

Benchmark code:

```elisp
(require 'benchmark)
(defun idris-delete-source-highlight-overlays (buffer)
  (with-current-buffer buffer
    (mapc 'delete-overlay (seq-filter
                           (lambda (overlay)
                             (overlay-get overlay 'idris-source-highlight))
                           (overlays-in (point-min) (point-max))))))

(defun load-file-bench (b1 b2)
  (idris-delete-source-highlight-overlays b1)
  (idris-delete-source-highlight-overlays b2)
  (with-current-buffer b1
    (idris-load-file))
  (with-current-buffer b2
    (idris-load-file)))

;; to force change Idris working directory and reload the files 
;;we test load continuously 2 files in different dirs
(defun idris-run-file-load-benchmark (n f1 f2)
  (let ((b1 (find-file-noselect f1))
        (b2 (find-file-noselect f2)))
    (idris-delete-source-highlight-overlays b1)
    (idris-delete-source-highlight-overlays b2)
    (garbage-collect)
    (let ((idris-semantic-source-highlighting-max-buffer-size 9000))
      (benchmark-run-compiled n (load-file-bench b1 b2)))))

(idris-run-file-load-benchmark
 10
 "/ncurses-idris/examples/snake/Main.idr"
 "/ncurses-idris/examples/control_pumpkin/Main.idr")
```